refactor Clock as a view component and no more as a controller
